bugfix, docs, info views, foreign pointer
But the foreign pointer is a thorny thing. It would work if there was
only one original source for references to a module or sample (when they
are loaded), but libmikmod allows the user to query the "current" module
at any time, replicating a foreign pointer. This is the only place in
the api this happens. We could ignore this by just not exporting that
method but ...

Inside the module you'll find a bunch of sample pointers. These samples
will presumably be freed if the module is freed, so you dont wan't foreign
pointers to these samples at all. This is similar to the pointer to module
from the current playing call.

Looked at weak references, not applicable.

We might just have to drop the ForeignPtr aspect of ModuleHandles and
SampleHandles.