FFI fixes for MReader.
First, all the FFI calls are now imported safe rather than unsafe.
The code for the Handle and ByteString reader wrappers was fixed.
This included a fix to the backwards reading of the return value of the
Read call. Documentation updated.

The test code playerLoadGeneric seems to work with or without the
-threaded runtime.