Updated several things, mainly MReader related.
The reader now needs to produce a bytestring rather than write a number
of bytes to a memory location pointed to be a pointer argument!

No more confusing Bools in the reader or use of C's EOf integer value.

The MReader wrapper for bytestrings is no longer an IO action. If you
happen to have an IORef you can make the reader is pure code! Now you
know that creating a bytestring reader has undocumented no side effects.

Minor doc fixes.